requirements.txt
Django==5.0.6
django_mysql==4.12.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
pandas==2.2.2
PyJWT==2.8.0
PyJWT==2.8.0
Unidecode==1.3.8
django-cors-headers==4.0.0
mysqlclient==2.1.1
python-dotenv==1.0.0
dj_database_url==2.1.0